fixed potential error with size factors and sparse data
