+++ batchglm/models/base_glm/input.py
@@ -211,7 +211,11 @@ class InputData(_InputData_Base):
 
     @property
     def size_factors(self):
-        return self.data.coords.get("size_factors")
+        if isinstance(self.data, SparseXArrayDataSet):
+            return self.data.size_factors
+        else:
+            return self.data.coords.get("size_factors")
+        #return self.data.coords.get("size_factors")
 
     @size_factors.setter
     def size_factors(self, data):
