+++ batchglm/unit_test/glm_all/test_acc_analytic_glm_all.py
@@ -1,18 +1,21 @@
+import logging
 from typing import List
 import unittest
-import logging
+import numpy as np
 import scipy.sparse
 
 import batchglm.api as glm
-from batchglm.models.base_glm import _Estimator_GLM
-
-from .external import Test_AccuracyAnalytic_GLM, _Test_AccuracyAnalytic_GLM_Estim
+from batchglm.models.base_glm import _Estimator_GLM, _Simulator_GLM
 
 glm.setup_logging(verbosity="WARNING", stream="STDOUT")
 logger = logging.getLogger(__name__)
 
 
-class _Test_AccuracyAnalytic_GLM_ALL_Estim(_Test_AccuracyAnalytic_GLM_Estim):
+class _Test_AccuracyAnalytic_GLM_ALL_Estim():
+
+    estimator: _Estimator_GLM
+    simulator: _Simulator_GLM
+    noise_model: str
 
     def __init__(
             self,
@@ -23,6 +26,9 @@ class _Test_AccuracyAnalytic_GLM_ALL_Estim(_Test_AccuracyAnalytic_GLM_Estim):
             init_a,
             init_b
     ):
+        self.simulator = simulator
+        self.noise_model = noise_model
+
         if noise_model is None:
             raise ValueError("noise_model is None")
         else:
@@ -51,7 +57,7 @@ class _Test_AccuracyAnalytic_GLM_ALL_Estim(_Test_AccuracyAnalytic_GLM_Estim):
                 design_scale=simulator.input_data.design_scale
             )
 
-        estimator = Estimator(
+        self.estimator = Estimator(
             input_data=input_data,
             batch_size=batch_size,
             quick_scale=not train_scale,
@@ -60,14 +66,93 @@ class _Test_AccuracyAnalytic_GLM_ALL_Estim(_Test_AccuracyAnalytic_GLM_Estim):
             init_a=init_a,
             init_b=init_b
         )
-        super().__init__(
-            estimator=estimator,
-            simulator=simulator
-        )
+
+    def estimate(self):
+        self.estimator.initialize()
+        self.estimator.train_sequence(training_strategy=[
+            {
+                "learning_rate": 1,
+                "convergence_criteria": "all_converged_ll",
+                "stopping_criteria": 1e-6,
+                "use_batching": False,
+                "optim_algo": "irls_gd_tr"
+            },
+        ])
+
+    def eval_estimation_a(
+            self,
+            estimator_store,
+            init_a,
+    ):
+        if self.noise_model is None:
+            raise ValueError("noise_model is None")
+        else:
+            if self.noise_model == "nb":
+                threshold_dev = 1e-2
+                threshold_std = 1e-1
+            elif self.noise_model=="norm":
+                threshold_dev = 1e-2
+                threshold_std = 1e-1
+            else:
+                raise ValueError("noise_model not recognized")
+
+        if init_a == "standard":
+            mean_dev = np.mean(estimator_store.a[0, :] - self.simulator.a[0, :])
+            std_dev = np.std(estimator_store.a[0, :] - self.simulator.a[0, :])
+        elif init_a == "closed_form":
+            mean_dev = np.mean(estimator_store.a - self.simulator.a)
+            std_dev = np.std(estimator_store.a - self.simulator.a)
+        else:
+            assert False
+
+        logging.getLogger("batchglm").info("mean_dev_a %f" % mean_dev)
+        logging.getLogger("batchglm").info("std_dev_a %f" % std_dev)
+
+        if np.abs(mean_dev) < threshold_dev and \
+                std_dev < threshold_std:
+            return True
+        else:
+            return False
+
+    def eval_estimation_b(
+            self,
+            estimator_store,
+            init_b
+    ):
+        if self.noise_model is None:
+            raise ValueError("noise_model is None")
+        else:
+            if self.noise_model == "nb":
+                threshold_dev = 1e-2
+                threshold_std = 1e-1
+            elif self.noise_model == "norm":
+                threshold_dev = 1e-2
+                threshold_std = 1e-1
+            else:
+                raise ValueError("noise_model not recognized")
+
+        if init_b == "standard":
+            mean_dev = np.mean(estimator_store.b[0, :] - self.simulator.b[0, :])
+            std_dev = np.std(estimator_store.b[0, :] - self.simulator.b[0, :])
+        elif init_b == "closed_form":
+            mean_dev = np.mean(estimator_store.b - self.simulator.b)
+            std_dev = np.std(estimator_store.b - self.simulator.b)
+        else:
+            assert False
+
+        logging.getLogger("batchglm").info("mean_dev_b %f" % mean_dev)
+        logging.getLogger("batchglm").info("std_dev_b %f" % std_dev)
+
+        if np.abs(mean_dev) < threshold_dev and \
+                std_dev < threshold_std:
+            return True
+        else:
+            return False
+
+
 
 
 class Test_AccuracyAnalytic_GLM_ALL(
-    Test_AccuracyAnalytic_GLM,
     unittest.TestCase
 ):
     noise_model: str
@@ -99,17 +184,61 @@ class Test_AccuracyAnalytic_GLM_ALL(
             init_b=init_b
         )
 
-    def _test_a_closed_b_closed(self, sparse):
-        self._test_a_and_b(sparse=sparse, init_a="closed_form", init_b="closed_form")
+    def simulate_complex(self):
+        self.sim = self.get_simulator()
+        self.sim.generate_sample_description(num_batches=1, num_conditions=2)
+        self.sim.generate_params(
+            rand_fn_ave=lambda shape: np.random.uniform(1e5, 2 * 1e5, shape),
+            rand_fn_loc=lambda shape: np.random.uniform(1, 3, shape),
+            rand_fn_scale=lambda shape: np.random.uniform(1, 3, shape)
+        )
+        self.sim.generate_data()
 
-    def _test_a_closed_b_standard(self, sparse):
-        self._test_a_and_b(sparse=sparse, init_a="closed_form", init_b="standard")
+    def simulate_easy(self):
+        self.sim = self.get_simulator()
+        self.sim.generate_sample_description(num_batches=1, num_conditions=2)
 
-    def _test_a_standard_b_closed(self, sparse):
-        self._test_a_and_b(sparse=sparse, init_a="standard", init_b="closed_form")
+        def rand_fn_standard(shape):
+            theta = np.ones(shape)
+            theta[0, :] = np.random.uniform(5, 20, shape[1])
+            return theta
 
-    def _test_a_standard_b_standard(self, sparse):
-        self._test_a_and_b(sparse=sparse, init_a="standard", init_b="standard")
+        self.sim.generate_params(
+            rand_fn_ave=lambda shape: np.random.uniform(1e5, 2 * 1e5, shape),
+            rand_fn_loc=lambda shape: np.ones(shape),
+            rand_fn_scale=lambda shape: rand_fn_standard(shape)
+        )
+        self.sim.generate_data()
+
+    def setUp(self):
+        self._estims = []
+
+    def tearDown(self):
+        for e in self._estims:
+            e.estimator.close_session()
+
+    def _test_a_and_b(self, sparse, init_a, init_b):
+        estimator = self.get_estimator(
+            train_scale=False,
+            sparse=sparse,
+            init_a=init_a,
+            init_b=init_b
+        )
+        estimator.estimate()
+        estimator_store = estimator.estimator.finalize()
+        self._estims.append(estimator)
+        success = estimator.eval_estimation_a(
+            estimator_store=estimator_store,
+            init_a=init_a,
+
+        )
+        assert success, "estimation for a_model was inaccurate"
+        success = estimator.eval_estimation_b(
+            estimator_store=estimator_store,
+            init_b=init_b
+        )
+        assert success, "estimation for b_model was inaccurate"
+        return True
 
 
 class Test_AccuracyAnalytic_GLM_NB(
@@ -121,14 +250,14 @@ class Test_AccuracyAnalytic_GLM_NB(
     """
 
     def test_a_closed_b_closed(self):
-        logging.getLogger("tensorflow").setLevel(logging.ERROR)
+        logging.getLogger("tensorflow").setLevel(logging.ERROR),
         logging.getLogger("batchglm").setLevel(logging.INFO)
         logger.error("Test_AccuracyAnalytic_GLM_NB.test_a_closed_b_closed()")
 
         self.noise_model = "nb"
         self.simulate_complex()
-        self._test_a_closed_b_closed(sparse=False)
-        self._test_a_closed_b_closed(sparse=True)
+        self._test_a_and_b(sparse=False, init_a="closed_form", init_b="closed_form")
+        self._test_a_and_b(sparse=True, init_a="closed_form", init_b="closed_form")
 
     def test_a_standard_b_standard(self):
         logging.getLogger("tensorflow").setLevel(logging.ERROR)
@@ -136,9 +265,9 @@ class Test_AccuracyAnalytic_GLM_NB(
         logger.error("Test_AccuracyAnalytic_GLM_NB.test_a_standard_b_standard()")
 
         self.noise_model = "nb"
-        self.simulate_a_b_easy()
-        self._test_a_standard_b_standard(sparse=False)
-        self._test_a_standard_b_standard(sparse=True)
+        self.simulate_easy()
+        self._test_a_and_b(sparse=False, init_a="standard", init_b="standard")
+        self._test_a_and_b(sparse=True, init_a="standard", init_b="standard")
 
 
 class Test_AccuracyAnalytic_GLM_NORM(
@@ -156,8 +285,8 @@ class Test_AccuracyAnalytic_GLM_NORM(
 
         self.noise_model = "norm"
         self.simulate_complex()
-        self._test_a_closed_b_closed(sparse=False)
-        self._test_a_closed_b_closed(sparse=True)
+        self._test_a_and_b(sparse=False, init_a="closed_form", init_b="closed_form")
+        self._test_a_and_b(sparse=True, init_a="closed_form", init_b="closed_form")
 
     def test_a_standard_b_standard(self):
         logging.getLogger("tensorflow").setLevel(logging.ERROR)
@@ -165,9 +294,9 @@ class Test_AccuracyAnalytic_GLM_NORM(
         logger.error("Test_AccuracyAnalytic_GLM_NORM.test_a_standard_b_standard()")
 
         self.noise_model = "norm"
-        self.simulate_a_b_easy()
-        self._test_a_standard_b_standard(sparse=False)
-        self._test_a_standard_b_standard(sparse=True)
+        self.simulate_easy()
+        self._test_a_and_b(sparse=False, init_a="standard", init_b="standard")
+        self._test_a_and_b(sparse=True, init_a="standard", init_b="standard")
 
 
 if __name__ == '__main__':
